extract common functions to db helper
